Support repeating a set of ayat for #190, #347.
This change adds a range repeat count to AudioRequest, and enforces
it during playback time. In addition, it bundles an mp3 for the
basmalah and for isti3atha, especially for cases when repeat goes
across sura bounds.

In addition, this patch also:
- fixes the fact that sura tawba doesn't have isti3atha for most qaris
- probably sets range for "play from here" when multiple ayat are
  selected.

This patch allows us to build a ui for repeating a range of verses.