Better handling for when images don't load.
We now show a retry button and a text description if an image fails
to load due to no sdcard or no internet connection. This fix also
changes the dialog message for downloading data in an attempt to
convince users to download the images instead of dealing with a
potentially broken experience.

This commit only implements this on phone.