Extract page logic into a separate ViewGroup.
Introduce QuranPageLayout, which is responsible for the layout and
rendering of a page. Currently, this includes figuring out the page's
borders and background.

This solution allows code reuse between the various page types for
things that are constant. It also is more efficient, due to the fact
that the error layout is only inflated if it's needed, and due to
the fact that we've replaced a RelativeLayout with a FrameLayout. It
also takes us slightly closer to a fragment free architecture.