Replace 1920 width with 1260 on new installs.
In the existing code, devices with screen widths larger than 1280
would default to using 1920 images. This ended up being problematic
for 2 reasons:

1. there's a limit on the maximum width of an image that can be
loaded on Android. on older devices, this limit is 2048x2048, though
it's significantly higher on new devices.

2. performance - larger images require a larger GL texture, which
causes loading of images to stutter.

Consequently, on all new devices (without existing images), the
app will use 1260 images if the width is greater than 1280. While
these images are slightly lower quality than the previous ones, the
difference is barely noticable, even on a Nexus 6 in landscape mode,
but the performance difference (and the simplification of code that
results from this change) is very noticable.

Also took this opportunity to kill the splash screen.