Remove the LruCache for images.
We used to maintain a cache of the recently used images. In reality,
this was only useful if you moved backwards to the past few pages.
Since this is a rare use case, and since the memory usage is high,
and since the number of cached pages would typically be 2 or 3,
removed this cache for now.

In the future, we should introduce a pool of bitmap objects and use
inBitmap to load images.