Move translation update from QuranDownloadService.
In an initial attempt to cleanup the QuranDownloadService to support
"download all" and other things, wanted to pull out the translation
update code from QuranDownloadService. For now, this has been moved
to QuranActivity, where it will check in the background and notify
the user of new translations the next time around (if applicable).

In the future, we should handle this differently - instead of
showing a dialog requesting the user to visit the translations
screen, we can mark this in the translations view (since it may
make more sense).

Also updated okhttp to the latest version.