Allow the AudioManager to download audio.
In the interest of releasing soon and interating to improve, this
patch connects the audio manager to the existing blocking progress
dialog. While not ideal, it is quite likely that only minor fixes
will be made for now (the messaging of the dialog, adding an arabic
translation, marking done shuyookh as done). Improving the
implementation can be done in a future release insha'Allah.