Make QuranDataActivity a standard Activity.
There is a crash on a specific device that complains that the theme
is not an AppCompat theme (though it is, as defined in the manifest).
In any case, explicitly set it in the manifest and switched to just
being a standard Activity (since it really doesn't show a ui anyway,
except for the download related dialogs).