Fix parcelable crash when restoring SpinnerCompat.
For some reason, we were having issues restoring the state of the
Spinner when reviving the activity (BadParcelableException due to
a ClassNotFoundException).

After much searching and reading, solved the issue with this workaround,
which involves only having saved state in AbsSpinnerCompat instead of
having part of it in the parent and part in SpinnerCompat. This also
turns out to be what ActionBarSherlock's IcsSpinner did as well.

Note that this is a valid fix (though perhaps non-ideal due to
AbsSpinnerCompat not really being a true abstraction), and that, in
addition to fixing the crash, it does persist spinner state (including
the open state of the menu) for something like the JumpDialog. The
remaining spinners (shuyookh list, etc) that aren't persisting have
never persisted before. We should definitely fix this in the future,
as it's especially jarring when switching orientation on the device.