Attempt to work around notification crash.
On certain 4.4 devices, we have a crash with notifications. This is
an attempt to work around this crash - first, we use a different id
for the processing notification, and second, we avoid sending the
initial (unknown file count) notification during processing.

If this fixes it, great. If not, then if in fact all the crashes now
occur with the new notification id (for processing), we can just stop
showing those notifications on Kitkat as a work around for this crash,
(since they aren't critical and the user still gets notified if the
app is running).