Stop closing and re-opening databases.
One of the causes for unnecessary trouble with databases is that we
open and close databases again and again. This results in many problems,
especially since sometimes, we can't re-open a database, or we close it
while another thread is using it, etc. Instead, keep a cached instance
of the database helper per database and never close it.

We were already doing this with some of our helpers (bookmarks and the
database that knows the downloaded translations), but not with others
(namely audio, ayahinfo, and translations).