Update appcompat. Use dialog for Jump dialog.
Upgrading appcompat to v22 gives us the FixedListPopupWindow changes
automagically as part of AppCompatPopupWindow. Moreover, revert a
"fix" yesterday where application context was used instead of context
while initializing the DropdownPopup. While this fixed the jump issues,
it caused a set of other issues on other devices. Reverting this should
fix them, and using a dialog for 2.3, while not ideal, gets the job
done. Fortunately, 4.x devices do not have this problem.