Fix broken basmala. Fixes #529.
Introduce image patch 4. This patch set fixes the shadda on the basmala
for pages 597 and 598. Also, for many of the existing widths, it fixes
the sura header being too close to the suras for pages 1 and 2. For
some of the widths, it also fixes the relative size of some of the pages
as seen in pages like 213, 270, 565, 576.

This also moves the directory to be /patches/v[version]/ instead of just
putting everything directly in /patches/.

A future patch will update the original download zip to be correct as well,
since, until then, a user with no data will have to download the images zip,
followed by the patch.