Use new path for downloading images.
This new directory contains the newest image zips. For now, did not want
to change the existing zips in place, since this can cause incorrect
downloads (using part of an old file and continuing with part of the new
one). Eventually, this will be inevitable, since we're likely to symlink
the old files to the new ones at some point (so that the many users who
have not yet upgraded can still get the newest images if downloading
their images for the first time).

Also see #484 for more discussions around this problem.