Let audio buttons take the entire width on phone.
Instead of keeping each button a fixed width alone, instead stretch to
fill the entire available width for each button. This is only currently
applied to phone, since on tablet (ex a nexus 10), this is not ideal at
the moment. Ideally, on tablet, we should apply a limit to the amount of
space the bar can take up, and center and expand the items within that
space.