Fix toolbar being cut off pre-Kitkat.
On pre-Kitkat, requestLayout on the toolbar's parent to avoid having
the toolbar be cut off. This is because the fitSystemWindows insets
change pre-Kitkat when the toolbar is shown or hidden, and thus, if
we don't requestLayout on the toolbar, the height is incorrect, thus
causing the toolbar to not animate fully.