Update most dialogs to use material design style.
This updates most dialogs to use the appcompat AlertDialog. This patch
also improves the layouts on some of the dialogs, and brings up the
keyboard by default for "add tag" and "jump" dialogs.

This does not update any ProgressDialogs (since ProgressDialogs are not
mentioned in the material design guidelines and are considered "bad ui,"
appcompat will not support them - Quran should likely move away from
using these in the few places remaining within the app).

It also does not update the list preference for "download amount," but
this can easily be done later on insha'Allah.