Support building a Qaloon specific app. Refs #355.
Many people have requested various different sets of pages (Qaloon,
Warsh, and the Urdu Naskh pages are the biggest requests). This patch
makes it possible to build a separate app for Qaloon. Most of the code
is shared between the two apps, minus things that need to be extracted
for that specific build.

Moving forward, Qaloon (and Urdu, which someone else is working on at
the moment now), will live in the same directory as the standard Quran
does. These flavors will share the audio and translations (though we
can, and likely will, disable some of the audio on a ui level for some of
the flavors). The pages will live in a subdirectory under quran_android,
as will the ayahinfo file (currently, for Qaloon, the images are stored
in qaloon/, and the database is in qaloon/databases).