Bugfixes for width code changes.
Due to having forgotten to add the "_" before the tablet size in naskh,
made it so that adding the "_" is not the job of the PageProvider. In
the future, will likely make getWidthParam and getTabletWidthParam
return ints (the only issue currently with doing this is the override
parameter).

Also, when trying to copy the arabic search database from a different
type (qaloon, naskh, etc), attempt to make the directory before trying
to copy the database file (since, if that flavor is the first flavor
that the user is running, they will otherwise not have that directory).