Use qurancomplex uthmani hafs font for arabic text.
The various Arabic fonts on Android (and pretty much all standard Arabic
fonts) do not have all the glyphs present for the text present in the
translation view. Consequently, there were many cases where tashkeel was
improperly rendered, even though the data in the underlying database was
correct. This patch fixes this issue by using the uthmani hafs font from
qurancomplex to render the arabic in the translation view. Fixes #477.