Separate basmallah from first ayah in translation.
The sqlite database with the arabic data (quran.ar.db) has the basmallah
attached to the first ayah for all suras (with just a space separating
them). This code manually fixes this problem by removing the basmallah
string from the first ayah of all non-fatiha/non-tawba suras, and adding
it underneath the sura title header instead.

This is implemented safely, so that in the future, if we decide to fix
the problem at the database level, people with new databases and old
code should still see the right data. Fixes #543.