Separate font based on build. Use noorehira for naskh.
This change is primarily to allow using noorehira for rendering the
arabic text in the translation view on naskh. The fonts are now bundled
depending on the app type that is built and configured in the app
specific configuration. Added the noorehira font and credits for it in
the naskh specific build.