Remove old audio migration code.
Long ago, the app used to store audio data in /Android/data (pre-2.2.0,
which was released in 2012). In 2.2.0, we changed it such that the
default directory was /sdcard/quran_android, and, consequently, had a
migration script on startup to migrate the audio. This patch removes
that code, since it's unlikely that it will be hit nowadays.

Simplifying the startup / update code is essential for the permissions
work needed to target M.