Don't upgrade preferences on a new install.
Currently, PREF_VERSION keeps track of the last version of the app that
was installed, and can be used to change things between app launches.
After the last patch, this is now just used to determine whether or not
existing partial zip files should be removed or not, and there's no
reason to run this code on a brand new install, so don't.