Initial commit targetting android M. Bugfixes.
This changes Quran to compile against Android M. It also changes the way
directories are dealt with - if the install is most likely a new
install, data is saved in /sdcard/Android/data (which requires no
explicit permission from the user). Otherwise, the user is asked to
grant permission for writing the sdcard. Given that permission, Quran
can read the previous files (/sdcard/quran_android), otherwise falling
back to /sdcard/Android/data.

Also fixed a problem where, when a page couldn't be loaded due to the
sdcard not being mounted, for example, the error message showing the
problem would not appear.

It is probable that we'd want to give the user a setting to override
this (in case the user accidentally said no, then wanted to say yes
again) - this would likely live in settings, and would likely also have
an entry point from the initial download screen.

Refs quran/quran_android#574.