Restore default to /sdcard, even on M.
In the previous commit, made it so that M defaults to using external
files dir, which is writable without permission. Since audio downloads
are typically large and some people don't have easy access to the
internet, it'd be especially troublesome for them to uninstall the app,
only to find all their audio gone.

Consequently, this change makes M also default to /sdcard, thus asking
permission on first launch. If this permission is not granted, then
fallback to external files dir as usual. Also changed messaging to
clarify that not granting the permission would result in all the data
being removed when the app is uninstalled.