Only use uthmani font on api 18 and above.
There are some pretty bad Arabic font issues pre-api 18. These aren't
specific to the uthmani font, as they also happen with no explicit
typeface set (i.e. system Roboto font) and with the naskh font. The
issues are less apparent when using the standard font, however, and so
it seems like it may be a good idea to gate this to only be 18+.

Also use the old quran.ar.db in case we're falling back to the standard
font on the naskh version.

Not really convinced as to whether this is the right thing to do or not,
since the issue appears without an explicit typeface as well on 4.0.3.
However, the app has been that way for years (before this recent
change), and no one has ever complained. Installed an older version and
can confirm that this behavior used to exist before too.

See also quran/quran_android#580.