Temporarily disable automatic Android backups.
This is a workaround for a pretty unfortunate bug - as of Android M, backup
happens automatically for everything in the internal directory (most notably,
shared preferences). This is very problematic, because it results in the app
never prompting for permissions on M and assuming that the permission is
disabled due to the combination of the fact that the app thinks we showed
the preference dialog, but at the same time, the show rationale method tells
us not to show rationale (since we didn't request permissions yet).

More discussion in quran/quran_android#586.