Don't use SharedPreferences directly in QuranDataActivity.
Proxying through QuranSettings is important, since future work will
likely split up a group of settings into a different file. Having one
place for this code to tunnel through is likely to help avoiding bugs.
More work still needs to be done to use QuranSettings in other parts of
the app as well.