Fix bugs with permission.
Android's shouldShowRequestPermissionRationale only returns true if the
user has been prompted for permissions before and declined (assuming
they didn't select, "never ask me again"). In the previous code, we
always wanted to show rationale before showing the permissions dialog,
so we'd only show it if the rationale dialog hadn't been presented
before (or if shouldShowRequestPermissionRationale returned true).

This is wrong, since the user could say false to our dialog, causing
future launches to return false for both our check as to whether we
showed the dialog and shouldShowRequestPermissionRationale. This patch
fixes this by instead changing the setting to reflect when we showed the
Android permission dialog itself (i.e. called requestPermissions).

Note that part of the work for quran/quran_android#586 (fixing auto
backup) requires that this shared preference not be persisted during
backup. This should be in another patch insha'Allah.