Fix one more shared preferences crash.
2.6.7-p2 introduced a bad bug in which it mixed up an int and a string
used as error indications for bad downloads. As a result, most poeple
crashed on launch. 2.6.7-p3 was pushed and fixed this for 99% of people.
However, there is still a small set of people (84 people at the time of
this writing) who didn't crash at the bad logic in 2.6.7-p2, and thus
stored the error data in the wrong format. When 2.6.7-p3 fixed it, those
people started having crashes as well.

This should (insha'Allah) fix this issue.