Separate ui and model for import. Added tests!
This commit does 2 very important things for the new QuranImportActivity
flow - first, it separates the ui logic from the business logic. Second,
it adds testing libraries and Quran for Android's first set of unit
tests!

While the unit tests don't cover all of the cases, they are a start.
Insha'Allah in the future, hoping that the project gets many more tests.