Use Timber for logging.
This has two benefits - the first is that we don't get logging to adb by
default in production builds (which is bad practice). The second is that
we can install a specific production log listener that records the last
few lines, and, in the case of someone reporting a problem, we can
easily attach logs to know what's happening.