Initial work on combined bookmarks/tags screen.
This patch introduces a separate NewBookmarksFragment that can correctly
render bookmarks sorted by date or location in the Quran, and grouped
either by tags or by the existing behavior (page bookmarks, ayah
bookmarks). Note that the contextual action menu is not yet implemented in
this patch.

In order to test this, change BookmarksFragment.newInstance() to
NewBookmarksFragment.newInstance().