Improve performance of bookmark queries.
Previously, everytime we'd get a bookmark with tags (which now is the
only way we get bookmarks), we'd do one query bookmarks, and then, for
each bookmark, we'd do a second query to get tags. Needless to say, that
was both unnecessary and super expensive. Replaced this with a left join
instead to do one query instead of a query per bookmark.