Set project titles as untranslatable
Fixes https://github.com/quran/quran_android/issues/600.

Also includes minor punctuation changes.
Rule of thumb for using periods in preference summaries is that
it should not be used if the text consists of  only a word, phrase
or a sentence; if the summary text is more than a sentence, then
period(s) should be used where appropriate and at the end.