Fix tags being incorrectly checked.
This fixes a bug introduced in e389760e6a23f23057906a3ead6f54e7355734cf,
where isChecked was removed from Tag, consequently forcing
TagBookmarkDialog to have a list of tags and a list of checked tags. The
bug was that on refresh, the list of checked tags was never being
cleared.

In practical terms, this meant that if one unbookmarks a tagged ayah and
goes to see the ayah tags, they would see the previous tag(s) checked, even
though unbookmarking the ayah removes all tags.

This patch fixes this issue by clearing the list of checked tags upon
refresh (instead of only adding newly checked tags to the list).