Require unbind to take a type as a parameter.
This is to fix an edge case - when changing the arabic setting, what
ends up happening is that (assuming the activity doesn't get destroyed)
onResume calls finish() and relaunches QuranActivity. Consequently, the
old fragment gets onStart(), then the new fragment gets onStart(), and
then the old fragment gets onStop(). The old logic cleared the reference
in onStop(), but in this case, we're clearing the wrong reference (and
thus no connections work for the now attached fragment because the old
one caused it to be unattached).

There were two potential fixes for this - one is what this patch does
(check to make sure that what we're unbinding is what's actually bound,
and, if so, unbind). The other option is to use onPause/onResume instead
of onStop/onStart, but will stick to the first solution for now unless
it is deemed unreliable.