Fix QuranActivity page order on language change.
This is an existing bug affecting the current production version as
well. Essentially, if the activity gets killed when you go to change
the is Arabic flag, when you come back, the ViewPager will have its
pages in the wrong order.

This is due to the way that FragmentStatePagerAdapter restores
fragments. This fixes the problem by introducing an item id for each fragment
that does not change irrespective of Arabic or English. Consequently,
restore always restores the correct fragment for that position.

Note that one minor bug still remains for this case, which is that the page
does not always reset to the first page, despite the call in onCreate.