Fix a tags bug in the ayah ViewPager.
If we switched from tags to audio in the ayah ViewPager, because the
limit for that pager is the default, we end up stopping the tags
fragment and restarting it when we go back. Before this fix, restarting
it would cause the fragment to be empty, since refresh only happened
when one of setBookmarksMode or setAyahBookmarkMode are called on the
presenter. This fixes the problem by re-playing the last result onBind.