Initial implementation of tags under bookmarks.
Render the tags under each bookmark when not grouped by tags. This
implementation has a few current issues - first, we only show a fixed
number of tags depending on the available space (and show no indication
of additional tags if any exist). Second, the ordering of the tags differs
depending on RTL or LTR. Third, because the current height of the row items
is a fixed height, things are a bit cramped. Finally, we should look at
performance of rendering these row items.