Fix crash if Arabic database can't be opened.
If the Arabic database doesn't exist or is corrupted, the app will crash
on launch. This catches the exception and handles it gracefully. Note
that this is not the ideal solution, since this means that even if we
download the database later on that we still won't be able to open it
until we exit and restart the app.