Replace header with QuranHeaderView.
Replace the LinearLayout with weights, which causes the title to be
measured twice, with a custom view that only lets its child views be
measured once. Note that further optimization is possible and likely to
be added in a future patch, whereby the inner 2 views can be replaced
with simple draw commands and this can just be a View instead of a
ViewGroup.