Fix current page not being updated in bookmarks.
Fix the current page not being updated in bookmarks. Also depend on
RxPreferences to make listening for preference changes a bit easier.
The thought is that RxPreferences will allow us to simplify some code in
various places, but if it ends up not getting a lot of usage, we can just
create an Observable in QuranSettings which emits an event everytime the
current page is changed, for example.