Only show share dialog after export if relevant.
If relevant, we'll show the share dialog after exporting data so that a
person can save to Dropbox, a different place on the file system, or
send via email, etc. However, Quran itself can handle this intent. In
rare cases, nothing can handle the intent except Quran itself, so let's
not show the dialog in this case and instead Toast where the file was
saved.