Let travis-ci test release builds.
By default, travis runs ./gradlew test, which tests all combinations.
These were failing on release builds due to the crashlytics key being
missing. Ideally, however, we want to be able to test release build (so
that we can catch proguard issues early on, for example). This patch
introduces a gradle flag, disableCrashlytics, which, when building
release or beta, disables crashlytics.