Implement nicer looking notifications.
This patch draws a custom "album art" image for use in the media style.
It looks best on apis 21+, where the notification background color can
be changed, but looks fine pre-21 as well. Furthermore, since there's a
possibility that generating this image throws OOM, fall back to the
previous, non-MediaStyle notification if we can't generate the image.