Prompt for db download on switching to translation
When we renamed the database for naskh, lots of people who had the
database didn't get prompted for a download unless they long pressed and
asked to play audio. They then complain in reviews that there is no
arabic in translation view. This fixes the problem by prompting for the
download of the arabic database if it doesn't exist before switching to
translation view.