Cleanup code around MediaSession.
After the patch, we're not getting crashes any more - it seems as though
explicitly passing the ComponentName has fixed this crash (and that the
crash was that there was more than one possible intent that could handle
the media button initially). This consequently cleans up the code around
this again.