Move import method to BookmarkModel
The import method used to live in QuranImportPresenter, which made it
difficult to notify various consumers of bookmarks that they changed.
Consequently, moved the import method to BookmarkModel (which is where
it belongs anyway), and invalidate the cache after import. Moreover, add
some logging to help some people who are running into issues with import.
Fixes #633.