Update strings.xml for Indonesian language (#669)
Replaced several english words that are commonly used by indonesian users for corresponding indonesian language words that are equally commonly used or are getting popular, such as "Unduh" for Download and "Berkas" for File.

Rephrasing and redactional changes to current indonesian translation to reflect the more exact meaning of the english prompts and directions.

Re-order strings in indonesian translation to the same order as the english. Complete all string translations which were not translated in previous file.