Build against Nougat
This updates Quran to build against Android N, and updates the support
libraries. As a result of this, the Spinner implementation changed
again, thus breaking our mirrored version. This patch thus reimplements
SpinnerCompat as a subclass of the newest AppCompatSpinner, while
overriding the correct methods to make sure the dropdown width is
correct.