Better support for multi-window mode on Android N
Quran actually worked pretty well with multi-window mode, minus one
issue - toggling the ToolBar wasn't working. This is due to the fact
that in multi-window mode, "apps cannot hide the status bar."

This mostly fixes #679, with the exception of the fact that the status
bar still overlays the page in multi-window mode. While this itself can't
be fixed (because "apps cannot hide the status bar if they are not running
in full-screen mode"), it should be possible, in the future, to drop
Quran out of full screen mode when in multi-window mode.