Manually draw borders in QuranPageLayout
This patch replaces the ImageViews for the left and right borders with
drawables that are shared between pages and are just drawn on the
canvas.

This patch also switches QuranPageLayout to be a custom ViewGroup instead
of a subclass of FrameLayout, since later on, we'll want to move the
drawing of the page header and footer text here.