Make seekbar preference consistent with prefs
On Lollipop and above, the seekbar preference was misindented (and had
the incorrect text size) when compared to the remaining items. This
patch fixes this. In the future, these preferences should be replaced
with something else (i.e. to make it easier to see the effects of the
adjustment instead of just having a meaningless number for size or
brightness).