Improve header and footer text display
Previously, the header and footer were drawn completely according to
where the bounds of the page were (so as to not take extra space). This
is problematic in many places (especially in landscape), because ayah
headers aren't part of the database (and even if they were, they're at
the very top of the screen anyway) - consequently, the text overlays
them. This patch draws text at a fixed height (and pads the view to make
space for the text).