Fix explicit LTR hardcodes for lint
Most of these LTR hardcodes are needed for supporting pre-api 17. To fix
this, split out layout-ar and layout-ar-v17, with the latter being an
exact copy of the respective files from layout.

Some of the other cases (the various ViewPager indicators) actually
always receive data in the correct order, so they need to be LTR. These
cases have been placed in the code instead. It should be possible to
remove  some of these explicit layout direction attributes in the
future after some additional work.