Implement recent pages
This implements recent pages for Quran (using similar behavior to Quran
for iOS). A table of recent pages now exists within the bookmarks
database. Whenever a Quran page is selected, it is added to the table.
When the page changes, the new page replaces the old one in recents. The
recents table currently maintains a maximum of 3 items. Fixes #691.