Improve Recent pages implementation
Previously, recent pages was doing a database query every time a page
changed in order to update the recent pages table. This patch cleans
this up by batching update information and only writing to the database
in onPause.

It also provides a much cleaner implementation for getting the last page.
Previously, getting the last page meant getting recents and then
returning the last page. This implementation adds an Observable with the
very latest last page seen (even if it hasn't been committed to the
database yet). This observable is prepopulated with the most recent page
from the database and thereafter never depends on the database.

This patch also refactors recent pages out into a RecentPageModel and
adds unit tests for it. A RecentPagePresenter is also added to begin
the process of decoupling business logic from PagerActivity.