Fix broken unit test (#720)
The unit test called a method that called onNext on a subject. This
subject was subscribed to as an observable to do some extra work, but
was doing this work on an io thread. The bug was that the unit test was
doing the checks before the code that was supposed to run on the io
thread ran (since the unit test itself is synchronous unless we
explicitly wait for something). The fix is to explicitly wait for the
observable to get an event, thereby letting us know that the code was
executed, before checking that the right things were called.