Fix tags test by using a TestScheduler
The tags test used to set the main thread to be "Schedulers.immediate()"
so things set to run on it would run immediately. Since RxJava 2 removes
immediate() since it wasn't properly implemented, updating the test to
use the test scheduler so we can make sure the necessary work is done
before checking results.