Fix flakey recent page model test
The testUpdateLatestPageWithSlowRecents test was flakey - it is intended
to ensure that if somehow, a person got to a page before the data came
back from the database, that we'd disregard the database data in this
case and use the person's pages instead.

This change makes the test deterministic by artificially delaying the
recent pages result, and only advancing the time after a page is
emitted.