Share more code between Quran and Tablet
This diff moves most of the "ayah tracking" code (highlighting and
unhighlighting of ayahs, getting the ayah at a position, etc) into a
shared presenter to be used by both QuranPageFragment and
TabletFragment. The fragment has to specify an array of its pages, and
any command is then given to each page to handle it depending on its
page type.

This greatly simplifies the existing tablet code, which historically was
very similar to the phone code, with the exception of a check for which
page to map a command to. The fact that its shared means that Quran and
Tablet can now have similar experiences (without double the effort).