Support multiple pages in TranslationPresenter
This patch updates TranslationPresenter to be able to deal with multiple
pages. This allows tablet to use one TranslationPresenter that is shared
between both the left and right pages (instead of one for the left page
and one for the right one). This also uses Dagger to inject
TranslationPresenters wherever they are needed.