Invalidate NumberFormat when Locale changes
Fixes #746. Because the callback would happen after the locale changed
(but before the change was fully in effect), when switching from Arabic
back to the default language, the app would make a new NumberFormat
using the locale, which at that point, was still the Arabic locale. This
causes Arabic numbers in between non-Arabic text on the main page.