Fix play button not working
The bug was introduced in a8f2b2d742026a542b8751a622252e584e6e72b6,
where the "next page" ayah was changed to use the previous page's ayah.
This caused us to try to play a range of ayat with an end less than
start, which caused nothing to play at all. This fixes #801.