Fix play button not working
The bug was introduced in f7cb2397f447c452daf33257b71b96e641c580ba,
where the "next page" ayah was changed to use the previous page's ayah.
This caused us to try to play a range of ayat with an end less than
start, which caused nothing to play at all. This fixes #801.