Android implements the unicode guidelines for plurals: http://www.unicode.org/cldr/charts/29/supplemental/language_plural_rules.html#fa
Unfortunately this means for Farsi/Persian, we have only "one" and
"other", and "one" is also used for "zero" (which will be ignored if
included).  If we include a specific message for "one" under plurals
(e.g., to say "one tag deleted" instead of "%d tag deleted"), that
will mess up the situation for "zero".

If there is a chance that the plurals might be used for value "zero",
we cannot customize "one".  Therefore, I believe it is safer to only
include "other" to avoid future confusion.