Improve Farsi translations (#847)
* Improved Farsi translations:
1) Several incorrect translations were fixed: vertical vs horizontal, before vs after, etc.
2) Some translations did not sound right in the particular context they were used in.
3) Correct spelling rules (e.g., regarding use of ZWNJ) had not been observed.
4) Many translations did not make any sense, or the argument replacement had been placed at the wrong part of the sentence.
5) Almost all missing translatables are now added.

* Android implements the unicode guidelines for plurals:
http://www.unicode.org/cldr/charts/29/supplemental/language_plural_rules.html#fa

Unfortunately this means for Farsi/Persian, we have only "one" and
"other", and "one" is also used for "zero" (which will be ignored if
included).  If we include a specific message for "one" under plurals
(e.g., to say "one tag deleted" instead of "%d tag deleted"), that
will mess up the situation for "zero".

If there is a chance that the plurals might be used for value "zero",
we cannot customize "one".  Therefore, I believe it is safer to only
include "other" to avoid future confusion.