Fix spinners cutting off for Farsi
This is a hack to fix #849. In summary, there is something about how
Farsi text is measured when it's prefaced with Farsi numbers (the same
doesn't happen with Farsi text without numbers or with Farsi text with
Arabic numbers, for instance). This hack works around this problem by
explicitly updating the layout params for the Spinner to reflect the
calculated width its dropdowns should use. Consequently, this forces
children inflated by the adapter to use that exact width without trying
to calculate a width based on wrap_content.