Support sharing ayahs with translation
In translation mode, long pressing an ayah now shows a toolbar (similar
to the one in the images screen), but with the share menu. Choosing to
copy or share will copy the verse in Arabic (if it is not disabled in
the settings) followed by all currently enabled translations as text,
and then copy or share.