Fixing comments & Adding Features
This is a major overhaul of the code to comply with comments received
and add additional features such as going to specific pages both in
arabic and english. Currently this codes allows to go to suras in arabic
and english (only english translated names, also 2 suras are not
currently working Qaf and Jinn). You can change the voice commands
language as well.