Implement a data source for Quran data
Currently, Quran data is overridden per flavor in a not-so-clean way.
This moves towards having a QuranDataSource that's irrespective of
flavor. This helps move towards a world in which Quran for Android can
host multiple types of pages and images (potentially simultaneously).

These classes are not in use yet and will insha'Allah be used in future
patches to replace BaseQuranData and QuranData subclasses. Note that
some arrays were also removed from these data sources since they aren't
strictly necessary.