Replace QuranData classes with QuranDataSource
This change is important since it helps prepare Quran for Android to be
able to handle multiple page types at the same time. As far as data is
concerned, all that is needed is to pass the correct data source to
QuranInfo, and the same app can then handle different types of pages.