Data source should be acquired from PageProvider
It's a little too troublesome to have both a QuranDataSource and a
PageProvider (which have no direct tie in together). This patch makes
things a little easier by making QuranDataSource obtained via the
PageProvider itself.

This in and of itself is non-ideal for a few reasons (ex
QuranDataSourceProvider being public, and app depending on both common
and pageinfo modules), but this is a temporary solution that should go
away in the near future in sha' Allah.