Let page size calculators not require the display
Pass a DisplaySize (data class with an x and y) instead of a Display to
the various calculators. Moreover, switch all calculators to use
getRealSize above api 17 and getSize below (previously, getRealSize was
called reflectively and we'd fall back to getSize (due to it being @hide
before api 17)).

This patch also changes madani (and derivatives) to use getRealSize when
available instead of always just using getSize.