Start services as foreground when it makes sense
Change some of the startService calls to startForegroundService on Oreo
and above. These changes are mostly in PagerActivity, where one download
can complete before calling startService again to start another download
(or to start the audio playback). In most of the other cases, the app is
in the foreground when the service starts and so explicitly asking for a
foreground service isn't necessary in those cases.