Prepare implementation for new madani images
The last release of Quran for Android (which was rolled out to a very
small number of people before being halted) had a set of new Madani
images. This patch prepares for replacing these with similar images from
the Ayah application (with permission). The benefits of these images are
that they allow custom drawing of the ayah markers, and they have been
tested over a few years as they're used within the Ayah application. One
final benefit is that the file size for download is smaller.

This patch prepares Quran to be able to query the corrosponding data
from the ayahinfo database if it exists and bubble it back to the image
view for future drawing purposes.