Fix slow drawing of ayah and headers on new madani
When new madani was implemented, the code to get the ayah coordinates
and sura headers was bundled with the ayah fetching code. This code
would run after the page coordinates were loaded, which would run 500ms
after the load call.

This patch changes it such that the ayah marker coordinates and the sura
header coordinates are part of PageCoordinates. Instead of waiting 500ms,
these are now loaded right away. The ayah highlighting data is then
loaded after 500ms from the loading of the page coordinates.