Fix audio ayah highlighting in landscape
This patch fixes audio highlighting in landscape. Before this patch, the
page would constantly scroll (especially in the case of a large ayah).
This patch improves it so that scrolling only happens if 1. an ayah is
small enough to fit on the screen but isn't entirely on the screen, or
2. a long ayah is not already on the screen. This fixes #1010.