Update arabic databases to use uthmani text (#1031)
This patch updates the databases to use uthmani text. It renames the
database for madani (to remain backward compatible with previous
versions). It also updates the version of the Hafs font. For modern sdk
versions, use QuranComplex data and fonts for display, and Quran Foundation text for sharing (https://github.com/alquran-foundation/quran-text). For older sdk versions, use simplified QuranFoundation text for sharing and display.