Remove translation by filename upon upgrading (#1051)
Because the old id for ibn kathir downloads the old version of the tafseer
(without footnotes, etc) and the new id downloads the new one, the ids
are different. Therefore, when upgrading tafaseer, remove any matching
items by name before updating to avoid getting 2 copies of the same
tafseer (one for the old id and one for the new, both thinking their
file is the correct one).