Get more information about images disappearing (#1060)
Whenever the images disappear, figure out whether or not the person had
granted permissions for external storage or not, and what sdk version
they were using. This might help figure out what happens.

Also fix a bug where if the app location is explicitly set to null, the
app never tries to find another location to store the app's files.