Try to copy updated Arabic database from assets
Due to the upgrade done to the Arabic database (switching to a
completely separate one), people will have to download the updated
database or risk not being able to share, see the ayah text above pages,
and so on. Given that this affects everyone (and that there's
intentionally no fallback logic to the old database due to its text not
being uthmani text), support copying the database from assets if it
ships with the app.