Fix a race condition while updating tafaseer (#1077)
When a tafseer is upgraded, there is code to remove entries with the old
filename from the database in certain cases. In these cases, the remove
was asynchronous despite the update itself being async, causing the
entry to usually be removed from the database after adding it. This
fixes this bug.