Fix juz' number in toolbar and bookmarks
Some juz's, like juz's 7 and 11, start towards the end of the page. The
madani mushaf thus labels the page with the previous juz'. This was done
in a previous commit. This patch updates the toolbar (which was not
updated at the time), and fixes bookmarked ayahs to reflect the correct
juz' that they are part of. Fixes #1087.