Only set Playing state after MediaPlayer.start()
Previously, the audio state was being transitioned to Playing once
onPrepared was called. This is problematic, however, because certain
commands (pause, for example) are invalid until the start method on the
MediaPlayer instance is called. This patch fixes it to only set the
state to playing after start is called.